fix ci tests
